// src/app/interfaces/unidadMoral.ts
export interface unidadMoral {
  id: string;
  UEDuenoid: string;
  Ofcid: string;
  FechaRegistro: string;
  RNPA: string;
  RFC: string;
  RazonSocial: string;
  Email: string;
  Calle: string;
  NmExterior: string;
  NmInterior: string;
  CodigoPostal: string;
  Locid: string;
  NmPrincipal: string;
  TpNmPrincipal: string;
  NmSecundario: string;
  TpNmSecundario: string;
  IniOperaciones: string;
  ActivoEmbMayor: string;
  ActivoEmbMenor: string;
  CantidadPescadores: string;
  ActvAcuacultura: string;
  ActvPesca: string;
  DocRepresentanteLegal: string;
  DocActaConstitutiva: string;
  DocActaAsamblea: string;
    
}